+++ lib/features/bottomNavBar/presentation/Transactions/addToTransactions.dart
@@ -1,8 +1,6 @@
 
-
-
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:tai/features/bottomNavBar/presentation/Transactions/transactionsModel.dart';
+import 'package:tai/features/bottomNavBar/presentation/Transactions/domain/transactionsModel.dart';
 
 addToTransactions(String senderId,String senderName,String senderImage, String receiverId,String receiverName,String receiverImage, double amount,String time ) async {
       var db = FirebaseFirestore.instance;
