+++ lib/features/bottomNavBar/presentation/home/notifications_page.dart
@@ -1,0 +1,19 @@
+// import 'package:flutter/material.dart';
+// import 'package:flutter_riverpod/flutter_riverpod.dart';
+// import 'package:tai/features/authentication/presentation/user_controller.dart';
+
+// class Notifications extends ConsumerWidget {
+//   const Notifications({super.key});
+
+//   @override
+//   Widget build(BuildContext context, WidgetRef ref) {
+//     final state = ref.watch(userProfileProvider);
+//     return Center(
+//       child: Column(children: [
+//         Text(state.email!),
+//         Text(state.username!),
+//         Text(state.totalBalance!.toString()),
+//       ]),
+//     );
+//   }
+// }
