+++ pubspec.yaml
@@ -1,12 +1,12 @@
 name: tai
 description: A new Flutter project.
 
-publish_to: 'none' 
+publish_to: "none"
 
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.1.3 <4.0.0'
+  sdk: ">=3.1.3 <4.0.0"
 
 dependencies:
   flutter:
@@ -26,14 +26,13 @@ dependencies:
   flutterwave_standard: ^1.0.8
   http: ^1.1.0
   provider: ^6.1.1
+  flutter_riverpod: ^2.4.9
   
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  
   flutter_lints: ^3.0.1
-
 
 flutter:
   uses-material-design: true
@@ -47,5 +46,3 @@ flutter:
       fonts:
         - asset: assets/fonts/Ubuntu-Regular.ttf
         - asset: assets/fonts/Ubuntu-Bold.ttf
-        
-
